// human-resource-management-app/src/models/payrollPeriod.ts
export default interface PayrollPeriod {
    maDotTraLuong: string;
    tenDotTraLuong: string;
    thoiGian: Date,
    tongNhanVienCanPhaiTra: number;
    tongNhanVienDaTra: number;
    tongNhanVienConPhaiTra: number;
    tongTienCanPhaiTra: number;
    tongTienDaTra: number;
    tongTienConPhaiTra: number;
    nguoiTraLuong: string;
    ghiChu: string;
  }